+++ cmd/function_deploy.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"time"
 
 	"github.com/spf13/cobra"
 	"github.com/vertigobr/safira/pkg/config"
@@ -196,15 +197,23 @@ func addPluginInAnnotations(functionName, namespace, kubeconfig, envFlag string,
 	for pluginName, plugin := range stack.Functions[functionName].Plugins {
 		if len(plugin.Type) == 0 || plugin.Type == "service" {
 			if len(plugins) > 0 {
-				plugins = plugins + ", " + d.GetDeployName(stack, pluginName)
+				plugins = plugins + ", " + d.GetDeployName(stack, fmt.Sprintf("%s-%s", functionName, pluginName))
 			} else {
-				plugins = d.GetDeployName(stack, pluginName)
+				plugins = d.GetDeployName(stack, fmt.Sprintf("%s-%s", functionName, pluginName))
 			}
 		}
 	}
 
 	if len(stack.Functions[functionName].Plugins) > 0 && len(plugins) > 0 {
-		err = d.AddPluginInAnnotationsService(d.GetDeployName(stack, functionName), namespace, plugins, kubeconfig, verboseFlag)
+		for i := 0; i < 30; i++ {
+			err = d.AddPluginInAnnotationsService(d.GetDeployName(stack, functionName), namespace, plugins, kubeconfig, verboseFlag)
+			if err != nil {
+				time.Sleep(time.Second * 2)
+			} else {
+				break
+			}
+		}
+
 		if err != nil {
 			return err
 		}